[fix/#455] 코드 리뷰 반영
